@@ -11,11 +11,11 @@
      ],
 	"license": "MIT",
     "require": {
-        "php": "^5.6 || ^7.0",
+        "php": "^7.2",
         "zendframework/zend-mail": "2.10.*"
     },
 	"require-dev": {
-        "phpunit/phpunit": "5.7.27",
+        "phpunit/phpunit": "^8",
 		"google/apiclient": "^2.0",
 		"squizlabs/php_codesniffer": "^3.5"
     },
@@ -30,7 +30,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.1.x-dev"
+            "dev-master": "1.3.x-dev"
         }
     } 
 }
