@@ -1,63 +1,62 @@
 <?php
-namespace Anod\Gmail;
 /**
- * 
- * 
  * @author Alex Gavrishev <alex.gavrishev@gmail.com>
- *
  */
-class GmailTest extends \PHPUnit_Framework_TestCase {
+namespace Anod\Gmail;
 
-	/**
-	 * @dataProvider providerGetRawMessageUID
-	 * @param $response
-	 * @param $expected
-	 * @throws GmailException
-	 */
-	public function testGetRawMessageUID($response, $expected) {
-		$protocol = $this->getMock('\Zend\Mail\Protocol\Imap', array('requestAndResponse'));
-		
-		$protocol
-			->expects($this->once())
-			->method('requestAndResponse')
-			->will($this->returnValue($response))
-		;
-		
-		$gmail = new Gmail($protocol);
-		$actual = $gmail->getMessageDataRaw("53559");
-		
-		$this->assertEquals($expected, $actual);
-		
-	}
-	
-	public function providerGetRawMessageUID() {
-		
-		$body = "--BODY--";
-		$header = "--HEADER--";
-		
-		$response = array(
-			array(
-				"9","FETCH",
-				array(
-					"UID","53559","RFC822.TEXT",$body,
-					"RFC822.HEADER",$header,
-					"FLAGS",
-					array("\Seen")
-      			)
- 			)	
-		);
-		
-		$expected = array(
-			"UID" => "53559",
-			"RFC822.TEXT" => $body,
-			"RFC822.HEADER" => $header,
-			"FLAGS" => array("\Seen")
-		);
-		
-		return array(
-			array($response, $expected)
-		);
-	}
-	
+use PHPUnit\Framework\TestCase;
+
+class GmailTest extends TestCase
+{
+    /**
+     * @dataProvider providerGetRawMessageUID
+     * @param $response
+     * @param $expected
+     * @throws GmailException
+     */
+    public function testGetRawMessageUID($response, $expected)
+    {
+        $protocol = $this->createMock(\Zend\Mail\Protocol\Imap::class);
+        
+        $protocol
+            ->expects($this->once())
+            ->method('requestAndResponse')
+            ->will($this->returnValue($response))
+        ;
+        
+        /** @var \Zend\Mail\Protocol\Imap $protocol */
+        $gmail = new Gmail($protocol);
+        $actual = $gmail->getMessageDataRaw("53559");
+        
+        $this->assertEquals($expected, $actual);
+    }
+    
+    public function providerGetRawMessageUID()
+    {
+        $body = "--BODY--";
+        $header = "--HEADER--";
+        
+        $response = array(
+            array(
+                "9","FETCH",
+                array(
+                    "UID","53559","RFC822.TEXT",$body,
+                    "RFC822.HEADER",$header,
+                    "FLAGS",
+                    array("\Seen")
+                  )
+            )
+        );
+        
+        $expected = [
+            "UID" => "53559",
+            "RFC822.TEXT" => $body,
+            "RFC822.HEADER" => $header,
+            "FLAGS" => array("\Seen")
+        ];
+
+        return [
+            [$response, $expected]
+        ];
+    }
 }
-
